Move serializers and viewsets to separate files
