+++ platalbank_core/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .event import *
+from .transaction import *
+from .account import *
