+++ platalbank_core/views/__init__.py
@@ -1,0 +1,3 @@
+from .event import *
+from .transaction import *
+from .account import *
