+++ platalbank_django/urls.py
@@ -17,9 +17,9 @@ from django.conf.urls import include, url
 from django.contrib import admin
 from rest_framework import routers
 
-from platalbank_core.models.event import EventViewSet
-from platalbank_core.models.transaction import TransactionViewSet
-from platalbank_core.models.account import AccountViewSet
+from platalbank_core.views.event import EventViewSet
+from platalbank_core.views.transaction import TransactionViewSet
+from platalbank_core.views.account import AccountViewSet
 
 
 router = routers.DefaultRouter()
